.ado/pipelines/python/StorageAccount/requirements.txt
azure-storage-blob
azure-identity
azure-keyvault-secrets
